% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loss_run.R
\name{loss_run}
\alias{loss_run}
\title{loss_run}
\usage{
loss_run(val_date, trans_)
}
\arguments{
\item{val_date}{date the valuation date of the loss run.  Claim values from `trans`
will be values as of the `val_date`}

\item{trans_}{data frame of claims transactions}
}
\value{
data frame of claims (1 claim per row) valued as of the `val_date`
}
\description{
view losses as of a specific date
}
